Fix ABI issue when upgrading PG from 15.3 to 15.4+
A previous change (commit 8d29760) added a workaround for a PostgreSQL
bug related to table access methods that was fixed in PG 15.4. The
workaround does a build-time check for the PG version to apply the
fix, but this won't work across PG server upgrades.

To avoid raising an error when running an extension compiled on PG <
15.4 on PG 15.4+, add also a runtime check.