Prepare VectorAgg exec path to handle arrow slots
The VectorAgg exec loop reads tuples directly from a compressed
relation, thus bypassing the DecompressChunk child node. This won't
work with arrow slots, which are read via a table access method.

To make the VectorAgg exec code similar to the standard pattern of
reading slots from child nodes, code specific to decompressing batches
is moved out of the main VectorAgg exec loop so that the loop only
deals with the final compressed batch slot instead of the raw
compressed slot. The code is instead put in a "get_next_slot"
function, which is called from the loop.

Also move the code to initialize vectorized filters to its own
"init_vector_qual" function, since it is specific to compressed
batches.

With these two function interfaces, it is possible to provide
implementations of the functions for handling arrow slots.